# backend/apps/calendar/routes.py
"""Calendar plugin FastAPI routes."""

from datetime import datetime

from beanie import PydanticObjectId
from fastapi import APIRouter, Depends, HTTPException, Query

from apps.calendar.schemas import (
    CalendarActionResponse,
    CalendarCalendarRead,
    CalendarCreateCalendarRequest,
    CalendarCreateEventRequest,
    CalendarCreateRecurringRuleRequest,
    CalendarEventRead,
    CalendarRecurringRuleRead,
    CalendarScheduleTaskRequest,
    CalendarUpdateCalendarRequest,
    CalendarUpdateEventRequest,
)
from apps.calendar.service import calendar_service
from core.auth.dependencies import get_current_user
from core.models import WidgetPreference
from shared.preference_utils import (
    preference_to_schema,
    update_multiple_preferences,
)
from shared.schemas import PreferenceUpdate, WidgetManifestSchema, WidgetPreferenceSchema

router = APIRouter()


# ─── Widgets ──────────────────────────────────────────────────────────────────

@router.get("/widgets", response_model=list[WidgetManifestSchema])
async def list_widgets() -> list[WidgetManifestSchema]:
    from apps.calendar.manifest import calendar_manifest
    return calendar_manifest.widgets


# ─── Events ───────────────────────────────────────────────────────────────────

@router.get("/events", response_model=list[CalendarEventRead])
async def list_events(
    user_id: str = Depends(get_current_user),
    start: datetime | None = Query(None),
    end: datetime | None = Query(None),
    calendar_id: str | None = Query(None),
    limit: int = Query(100, le=500),
):
    """List events with optional filters."""
    return await calendar_service.list_events(user_id, start, end, calendar_id, limit)


@router.get("/events/search", response_model=list[CalendarEventRead])
async def search_events(
    q: str,
    user_id: str = Depends(get_current_user),
    limit: int = Query(20, le=100),
):
    """Search events by query."""
    return await calendar_service.search_events(user_id, q, limit)


@router.get("/events/{event_id}", response_model=CalendarEventRead)
async def get_event(event_id: str, user_id: str = Depends(get_current_user)):
    """Get single event."""
    event = await calendar_service.get_event(event_id, user_id)
    if not event:
        raise HTTPException(status_code=404, detail="Event not found")
    return event


@router.post("/events", response_model=CalendarEventRead)
async def create_event(
    request: CalendarCreateEventRequest,
    user_id: str = Depends(get_current_user),
):
    """Create new event."""
    try:
        return await calendar_service.create_event(
            user_id,
            request.title,
            request.start_datetime,
            request.end_datetime,
            request.calendar_id,
            request.description,
            request.location,
            request.is_all_day,
            request.type,
            request.task_id,
            request.color,
            request.reminders,
        )
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.patch("/events/{event_id}", response_model=CalendarEventRead)
async def update_event(
    event_id: str,
    request: CalendarUpdateEventRequest,
    user_id: str = Depends(get_current_user),
):
    """Update event."""
    try:
        kwargs = {}
        if request.title is not None:
            kwargs["title"] = request.title
        if request.start_datetime is not None:
            kwargs["start_datetime"] = request.start_datetime
        if request.end_datetime is not None:
            kwargs["end_datetime"] = request.end_datetime
        if request.calendar_id is not None:
            kwargs["calendar_id"] = request.calendar_id
        if request.description is not None:
            kwargs["description"] = request.description
        if request.location is not None:
            kwargs["location"] = request.location
        if request.color is not None:
            kwargs["color"] = request.color
        if request.reminders is not None:
            kwargs["reminders"] = request.reminders
        return await calendar_service.update_event(event_id, user_id, **kwargs)
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.delete("/events/{event_id}", response_model=CalendarActionResponse)
async def delete_event(event_id: str, user_id: str = Depends(get_current_user)):
    """Delete event."""
    try:
        return await calendar_service.delete_event(event_id, user_id)
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


# ─── Conflicts ────────────────────────────────────────────────────────────────

@router.get("/conflicts/check", response_model=list[CalendarEventRead])
async def check_conflicts(
    start: datetime,
    end: datetime,
    exclude_event_id: str | None = Query(None),
    user_id: str = Depends(get_current_user),
):
    """Check for conflicting events."""
    return await calendar_service.check_conflicts(user_id, start, end, exclude_event_id)


# ─── Calendars ──────────────────────────────────────────────────────────────

@router.get("/calendars", response_model=list[CalendarCalendarRead])
async def list_calendars(user_id: str = Depends(get_current_user)):
    """List all calendars."""
    return await calendar_service.list_calendars(user_id)


@router.post("/calendars", response_model=CalendarCalendarRead)
async def create_calendar(
    request: CalendarCreateCalendarRequest,
    user_id: str = Depends(get_current_user),
):
    """Create new calendar."""
    try:
        return await calendar_service.create_calendar(user_id, request.name, request.color)
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.patch("/calendars/{calendar_id}", response_model=CalendarCalendarRead)
async def update_calendar(
    calendar_id: str,
    request: CalendarUpdateCalendarRequest,
    user_id: str = Depends(get_current_user),
):
    """Update calendar."""
    try:
        kwargs = {}
        if request.name is not None:
            kwargs["name"] = request.name
        if request.color is not None:
            kwargs["color"] = request.color
        if request.is_visible is not None:
            kwargs["is_visible"] = request.is_visible
        if request.is_default is not None:
            kwargs["is_default"] = request.is_default
        return await calendar_service.update_calendar(calendar_id, user_id, **kwargs)
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.delete("/calendars/{calendar_id}", response_model=CalendarActionResponse)
async def delete_calendar(calendar_id: str, user_id: str = Depends(get_current_user)):
    """Delete calendar and all its events."""
    try:
        return await calendar_service.delete_calendar(calendar_id, user_id)
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


# ─── Recurring ────────────────────────────────────────────────────────────────

@router.post("/events/{event_id}/recurring", response_model=CalendarRecurringRuleRead)
async def create_recurring(
    event_id: str,
    request: CalendarCreateRecurringRuleRequest,
    user_id: str = Depends(get_current_user),
):
    """Make event recurring."""
    try:
        return await calendar_service.create_recurring_rule(
            user_id, event_id, request.frequency, request.interval,
            request.days_of_week, request.end_date, request.max_occurrences
        )
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


@router.get("/recurring", response_model=list[CalendarRecurringRuleRead])
async def list_recurring_rules(user_id: str = Depends(get_current_user)):
    """List recurring rules."""
    return await calendar_service.list_recurring_rules(user_id)


@router.patch("/recurring/{rule_id}/stop", response_model=CalendarRecurringRuleRead)
async def stop_recurring(
    rule_id: str,
    user_id: str = Depends(get_current_user),
):
    """Stop recurring rule."""
    try:
        return await calendar_service.stop_recurring_rule(rule_id, user_id)
    except ValueError as e:
        raise HTTPException(status_code=404, detail=str(e))


# ─── Todo Integration ─────────────────────────────────────────────────────────

@router.post("/tasks/{task_id}/schedule", response_model=CalendarEventRead)
async def schedule_task(
    task_id: str,
    request: CalendarScheduleTaskRequest,
    user_id: str = Depends(get_current_user),
):
    """Schedule Todo task as time-blocked event."""
    try:
        return await calendar_service.schedule_task(
            user_id, task_id, request.start_datetime, request.duration_minutes, request.calendar_id
        )
    except ValueError as e:
        raise HTTPException(status_code=400, detail=str(e))


# ─── Preferences ──────────────────────────────────────────────────────────────

@router.get("/preferences")
async def get_preferences(
    user_id: str = Depends(get_current_user),
) -> list[WidgetPreferenceSchema]:
    prefs = await WidgetPreference.find(
        WidgetPreference.user_id == PydanticObjectId(user_id),
        WidgetPreference.app_id == "calendar",
    ).to_list()
    return [preference_to_schema(p) for p in prefs]


@router.put("/preferences")
async def update_preferences(
    updates: list[PreferenceUpdate],
    user_id: str = Depends(get_current_user),
) -> list[WidgetPreferenceSchema]:
    await update_multiple_preferences(user_id, updates, "calendar")
    return await get_preferences(user_id)